printer/requirements.txt
asgiref==3.4.1
attrs==21.2.0
brother-ql==0.9.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
future==0.18.2
h11==0.12.0
idna==3.3
packbits==0.6
Pillow==8.4.0
protobuf==3.19.1
pyusb==1.2.1
requests==2.26.0
structlog==21.5.0
twirp==0.0.5
urllib3==1.26.7
uvicorn==0.16.0
